+++ starlette/templating.py
@@ -5,7 +5,7 @@ from os import PathLike
 from starlette.background import BackgroundTask
 from starlette.datastructures import URL
 from starlette.requests import Request
-from starlette.responses import Response
+from starlette.responses import HTMLResponse
 from starlette.types import Receive, Scope, Send
 
 try:
@@ -23,9 +23,7 @@ except ModuleNotFoundError:  # pragma: nocover
     jinja2 = None  # type: ignore[assignment]
 
 
-class _TemplateResponse(Response):
-    media_type = "text/html"
-
+class _TemplateResponse(HTMLResponse):
     def __init__(
         self,
         template: typing.Any,
@@ -109,11 +107,20 @@ class Jinja2Templates:
         elif env is not None:
             self.env = env
 
+        self._setup_env_defaults(self.env)
+
     def _create_env(
         self,
         directory: "typing.Union[str, PathLike[typing.AnyStr], typing.Sequence[typing.Union[str, PathLike[typing.AnyStr]]]]",  # noqa: E501
         **env_options: typing.Any,
     ) -> "jinja2.Environment":
+        loader = jinja2.FileSystemLoader(directory)
+        env_options.setdefault("loader", loader)
+        env_options.setdefault("autoescape", True)
+
+        return jinja2.Environment(**env_options)
+
+    def _setup_env_defaults(self, env: "jinja2.Environment") -> None:
         @pass_context
         def url_for(
             context: typing.Dict[str, typing.Any],
@@ -124,13 +131,7 @@ class Jinja2Templates:
             request: Request = context["request"]
             return request.url_for(name, **path_params)
 
-        loader = jinja2.FileSystemLoader(directory)
-        env_options.setdefault("loader", loader)
-        env_options.setdefault("autoescape", True)
-
-        env = jinja2.Environment(**env_options)
-        env.globals["url_for"] = url_for
-        return env
+        env.globals.setdefault("url_for", url_for)
 
     def get_template(self, name: str) -> "jinja2.Template":
         return self.env.get_template(name)